# adapters/handlers/appointment.handler.go
package handlers

import (
	"github.com/labstack/echo/v4"
)

// CreateAppointment creates a new appointment for a diagnostic centre
// @Summary Create appointment
// @Description Create a new appointment for a diagnostic test
// @Tags Appointments
// @Accept json
// @Produce json
// @Param RequestBody body domain.CreateAppointmentDTO true "Appointment details"
// @Success 201 {object} handlers.AppointmentSwagger "SUCCESS_RESPONSE"
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/appointments [post]
func (h *HTTPHandler) CreateAppointment(c echo.Context) error {
	// Validation happens in middleware
	return h.service.CreateAppointment(c)
}

// GetAppointment retrieves an appointment by ID
// @Summary Get appointment
// @Description Get an appointment by its ID
// @Tags Appointments
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param Authorization header string true "Bearer token"
// @Param appointment_id path string true "Appointment ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Success 200 {object} handlers.AppointmentSwagger "SUCCESS_RESPONSE"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/appointments/{appointment_id} [get]
func (h *HTTPHandler) GetAppointment(c echo.Context) error {
	// Validation happens in middleware
	return h.service.GetAppointment(c)
}

// ListAppointments lists appointments based on filters
// @Summary List appointments
// @Description List appointments with optional filters
// @Tags Appointments
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param Authorization header string true "Bearer token"
// @Param diagnostic_centre_id query string false "Filter by diagnostic centre ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Param status query string false "Filter by status" Enums(pending, confirmed, in_progress, completed, cancelled, rescheduled)
// @Param from_date query string false "Start date (RFC3339)"
// @Param to_date query string false "End date (RFC3339)"
// @Param page query int false "Page number" default(1)
// @Param page_size query int false "Items per page" default(50)
// @Success 200 {array} handlers.AppointmentSwagger "SUCCESS_RESPONSE"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/appointments [get]
func (h *HTTPHandler) ListAppointments(c echo.Context) error {
	// Validation happens in middleware
	return h.service.ListAppointments(c)
}

// UpdateAppointment updates an existing appointment
// @Summary Update appointment
// @Description Update an existing appointment's details
// @Tags Appointments
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param Authorization header string true "Bearer token"
// @Param appointment_id path string true "Appointment ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Param RequestBody body domain.UpdateAppointmentDTO true "Updated appointment details"
// @Success 200 {object} handlers.AppointmentSwagger "SUCCESS_RESPONSE"
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/appointments/{appointment_id} [put]
func (h *HTTPHandler) UpdateAppointment(c echo.Context) error {
	// Validation happens in middleware
	return nil
}

// CancelAppointment cancels an existing appointment
// @Summary Cancel appointment
// @Description Cancel an existing appointment
// @Tags Appointments
// @Accept json
// @Produce json
// @Param appointment_id path string true "Appointment ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Param RequestBody body domain.CancelAppointmentDTO true "Cancellation details"
// @Success 200 {object} map[string]string "Cancellation success message"
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/appointments/{appointment_id}/cancel [post]
func (h *HTTPHandler) CancelAppointment(c echo.Context) error {
	// Validation happens in middleware
	return h.service.CancelAppointment(c)
}

// RescheduleAppointment reschedules an existing appointment
// @Summary Reschedule appointment
// @Description Reschedule an existing appointment to a new time
// @Tags Appointments
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param Authorization header string true "Bearer token"
// @Param appointment_id path string true "Appointment ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Param RequestBody body domain.RescheduleAppointmentDTO true "Rescheduling details"
// @Success 200 {object} handlers.AppointmentSwagger "SUCCESS_RESPONSE"
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 401 {object} handlers.UNAUTHORIZED_ERROR "UNAUTHORIZED_ERROR"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/appointments/{appointment_id}/reschedule [post]
func (h *HTTPHandler) RescheduleAppointment(c echo.Context) error {
	// Validation happens in middleware
	return h.service.RescheduleAppointment(c)
}

// ConfirmAppointment confirms an appointment
// @Summary Confirm appointment
// @Description Confirm an appointment by its ID
// @Tags Appointments
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param Authorization header string true "Bearer token"
// @Param appointment_id path string true "Appointment ID" default(123e4567-e89b-12d3-a456-426614174000)
// @Success 200 {object} map[string]string "Appointment confirmed successfully"
// @Failure 400 {object} handlers.BAD_REQUEST "BAD_REQUEST"
// @Failure 404 {object} handlers.NOT_FOUND_ERROR "NOT_FOUND_ERROR"
// @Failure 500 {object} handlers.INTERNAL_SERVER_ERROR "INTERNAL_SERVER_ERROR"
// @Router /v1/appointments/{appointment_id}/confirm [post]
func (h *HTTPHandler) ConfirmAppointment(c echo.Context) error {
	return h.service.ConfirmAppointment(c)
}