# swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Diagnoxix API",
        "title": "Diagnoxix",
        "contact": {},
        "version": "1.0"
    },
    "host": "diagnoxix.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/v1/appointments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List appointments with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "List appointments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Filter by diagnostic centre ID",
                        "name": "diagnostic_centre_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "pending",
                            "confirmed",
                            "in_progress",
                            "completed",
                            "cancelled",
                            "rescheduled"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS_RESPONSE",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.AppointmentSwagger"
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new appointment for a diagnostic test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Create appointment",
                "parameters": [
                    {
                        "description": "Appointment details",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateAppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "SUCCESS_RESPONSE",
                        "schema": {
                            "$ref": "#/definitions/handlers.AppointmentSwagger"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/appointments/{appointment_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get an appointment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Appointment ID",
                        "name": "appointment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS_RESPONSE",
                        "schema": {
                            "$ref": "#/definitions/handlers.AppointmentSwagger"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing appointment's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Update appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Appointment ID",
                        "name": "appointment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated appointment details",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateAppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS_RESPONSE",
                        "schema": {
                            "$ref": "#/definitions/handlers.AppointmentSwagger"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/appointments/{appointment_id}/cancel": {
            "post": {
                "description": "Cancel an existing appointment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Cancel appointment",
                "parameters": [
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Appointment ID",
                        "name": "appointment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cancellation details",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CancelAppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancellation success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/appointments/{appointment_id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Confirm an appointment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Confirm appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Appointment ID",
                        "name": "appointment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment confirmed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
        "/v1/appointments/{appointment_id}/reschedule": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reschedule an existing appointment to a new time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Reschedule appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Appointment ID",
                        "name": "appointment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rescheduling details",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RescheduleAppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS_RESPONSE",
                        "schema": {
                            "$ref": "#/definitions/handlers.AppointmentSwagger"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/handlers.BAD_REQUEST"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.UNAUTHORIZED_ERROR"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.NOT_FOUND_ERROR"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/handlers.INTERNAL_SERVER_ERROR"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CancelAppointmentDTO": {
            "type": "object",
            "required": [
                "appointmentID",
                "reason"
            ],
            "properties": {
                "appointmentID": {
                    "type": "string"
                },
                "reason": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "domain.CreateAppointmentDTO": {
            "type": "object",
            "required": [
                "amount",
                "appointment_date",
                "diagnostic_centre_id",
                "test_type"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "appointment_date": {
                    "type": "string"
                },
                "diagnostic_centre_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string",
                    "maxLength": 500
                },
                "payment_provider": {
                    "type": "string",
                    "enum": [
                        "PAYSTACK",
                        "FLUTTERWAVE",
                        "STRIPE",
                        "MONNIFY"
                    ]
                },
                "preferred_doctor": {
                    "type": "string",
                    "enum": [
                        "Male",
                        "Female"
                    ]
                },
                "test_type": {
                    "type": "string",
                    "enum": [
                        "BLOOD_TEST",
                        "URINE_TEST",
                        "X_RAY",
                        "MRI",
                        "CT_SCAN",
                        "ULTRASOUND",
                        "ECG",
                        "EEG",
                        "BIOPSY",
                        "SKIN_TEST",
                        "IMMUNOLOGY_TEST",
                        "HORMONE_TEST",
                        "VIRAL_TEST",
                        "BACTERIAL_TEST",
                        "PARASITIC_TEST",
                        "FUNGAL_TEST",
                        "MOLECULAR_TEST",
                        "TOXICOLOGY_TEST",
                        "ECHO",
                        "COVID_19_TEST",
                        "BLOOD_SUGAR_TEST",
                        "LIPID_PROFILE",
                        "HEMOGLOBIN_TEST",
                        "THYROID_TEST",
                        "LIVER_FUNCTION_TEST",
                        "KIDNEY_FUNCTION_TEST",
                        "URIC_ACID_TEST",
                        "VITAMIN_D_TEST",
                        "VITAMIN_B12_TEST",
                        "HEMOGRAM",
                        "COMPLETE_BLOOD_COUNT",
                        "BLOOD_GROUPING",
                        "HEPATITIS_B_TEST",
                        "HEPATITIS_C_TEST",
                        "HIV_TEST",
                        "MALARIA_TEST",
                        "DENGUE_TEST",
                        "TYPHOID_TEST",
                        "COVID_19_ANTIBODY_TEST",
                        "COVID_19_RAPID_ANTIGEN_TEST",
                        "COVID_19_RT_PCR_TEST",
                        "PREGNANCY_TEST",
                        "ALLERGY_TEST",
                        "GENETIC_TEST",
                        "OTHER"
                    ]
                }
            }
        },
...
        "domain.UpdateAppointmentDTO": {
            "type": "object",
            "required": [
                "appointmentID",
                "status"
            ],
            "properties": {
                "appointmentID": {
                    "type": "string"
                },
                "notes": {
                    "type": "string",
                    "maxLength": 500
                },
                "status": {
                    "enum": [
                        "pending",
                        "confirmed",
                        "in_progress",
                        "completed",
                        "cancelled",
                        "rescheduled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.AppointmentStatus"
                        }
                    ]
                }
            }
        },
...
        "handlers.AppointmentSwagger": {
            "description": "Appointment response for Swagger",
            "type": "object",
            "properties": {
                "appointment_date": {
                    "description": "format: date-time",
                    "type": "string",
                    "example": "2025-06-26T21:00:00Z"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-06-26T20:00:00Z"
                },
                "diagnostic_centre_id": {
                    "type": "string",
                    "example": "dc-001"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "patient_id": {
                    "type": "string",
                    "example": "user-001"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-06-26T20:30:00Z"
                }
            }
        },
        "handlers.BAD_REQUEST": {
            "description": "BAD_REQUEST response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "BAD_REQUEST"
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid request"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.INTERNAL_SERVER_ERROR": {
            "description": "INTERNAL_SERVER_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "INTERNAL_SERVER_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Internal Server Error"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.NOT_FOUND_ERROR": {
            "description": "NOT_FOUND_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "NOT_FOUND_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Data not found"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.UNAUTHORIZED_ERROR": {
            "description": "UNAUTHORIZED_ERROR response for Swagger",
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "UNAUTHORIZED_ERROR"
                        },
                        "message": {
                            "type": "string",
                            "example": "Unauthorized to perform this operation"
                        }
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    }
}